+++ HISTORY.rst
@@ -12,6 +12,7 @@ Unreleased
   Accounting). And set country by default using company user connected.
 * DT-5: Apply same logic for sale tax calculation than on invoices
 * DT-17 : Add group to control visibility of archive button on client view
+* DT-6: Add early payment discount on customer invoices
 
 **Bugfixes**